+ stop buying when buy lower spec weapon

+++ src/game/Shop.ts
@@ -1,4 +1,4 @@
-import { Character } from "@prisma/client";
+import { Character, Equipment } from "@prisma/client";
 import ICharacterService, { IncludeUserAndEquipment } from "../interfaces/services/ICharacterService";
 import IEquipmentService from "../interfaces/services/IEquipmentService";
 
@@ -18,14 +18,20 @@ export default class Shop {
         
         if (!character || !this.isChracterHaveEnoughCoin(character, coin)) return;
         
-        if (character.equipment)
+        if (character.equipment) {
+            if (!this.isNewEquipmentBetter(character.equipment, coin)) return
             await this.characterService.removeEquipment(character.id);
+        }
 
         let newEquipment = await this.equipmentService.createEquipment(character, coin, Math.ceil(coin / 4));
         if (!newEquipment) return;
         
         await this.characterService.removeCoinFromCharacter(character.id, coin);
         return character
+    }
+
+    private isNewEquipmentBetter(oldEquipment: Equipment, coin: number): boolean {
+        return oldEquipment.atk < coin
     }
 
     protected isChracterHaveEnoughCoin(chracter: Character, requireCoin: number): boolean {
